+++ DockerCompose/Docker-Compose.yml
@@ -1,0 +1,34 @@
+version: '3'
+services:
+  app:
+     image: openjdk:11
+     container_name: demoapi
+     ports:
+       - 8080:8080
+     tty: true
+     volumes:
+       - ../../demoapi:/srv:cached
+     working_dir: /srv
+     depends_on:
+       - userdb
+       - worlddb
+  worlddb:
+    container_name: worlddb
+    image: postgres:14
+    volumes:
+      - ./postgresql/world:/docker-entrypoint-initdb.d
+    environment:
+      POSTGRES_USER: hoge
+      POSTGRES_PASSWORD: hoge
+      POSTGRES_DB: worlddb
+    restart: always
+  userdb:
+    container_name: userdb
+    image: postgres:14
+    volumes:
+      - ./postgresql/user:/docker-entrypoint-initdb.d
+    environment:
+      POSTGRES_USER: hoge
+      POSTGRES_PASSWORD: hoge
+      POSTGRES_DB: usersdb
+    restart: always
