--- DockerCompose/mysql_user/Docker-Compose.yml
@@ -1,15 +1,0 @@
-version: "3.8"
-services:
-  mysql:
-    build: ./user/
-    volumes:
-      - ./user/db:/docker-entrypoint-initdb.d
-    image: mysql_user
-    ports:
-      - 3305:3305 #ホスト側ポート:コンテナ側ポート
-    expose:
-      - 3305
-    environment:
-      - MYSQL_ROOT_PASSWORD=hoge
-      - MYSQL_TCP_PORT=3305
-      - MYSQL_DATABASE=user
