--- DockerCompose/mysql_user/user/my.cnf
@@ -1,8 +1,0 @@
-[mysqld]
-character-set-server=utf8
-
-[mysql]
-default-character-set=utf8
-
-[client]
-default-character-set=utf8
