--- DockerCompose/mysql_world/Docker-Compose.yml
@@ -1,15 +1,0 @@
-version: "3.8"
-services:
-  mysql:
-    build: ./world/
-    volumes:
-      - ./world/db:/docker-entrypoint-initdb.d
-    image: mysql_world
-    ports:
-      - 3306:3306 #ホスト側ポート:コンテナ側ポート
-    expose:
-      - 3306
-    environment:
-      - MYSQL_ROOT_PASSWORD=hoge
-      - MYSQL_TCP_PORT=3306
-      - MYSQL_DATABASE=world
