+++ src/main/java/com/example/demo/repository/entity/users/ReqSelectUserCountryCodeEntity.java
@@ -1,0 +1,18 @@
+package com.example.demo.repository.entity.users;
+
+import java.util.List;
+
+import lombok.Getter;
+import lombok.Setter;
+
+/**
+ * selectUserCountryCodeエンティティクラス（リクエスト）
+ */
+@Getter
+@Setter
+public class ReqSelectUserCountryCodeEntity {
+	/*
+	 * 名前
+	 */
+	List<String> listName;
+}
