+++ src/main/resources/application.yml
@@ -3,15 +3,15 @@ spring:
   datasource:
     ## worldDB
     world:
-      driver-class-name: com.mysql.cj.jdbc.Driver
-      url: jdbc:mysql://127.0.0.1:3306/world?allowMultiQueries=true
-      username: root
+      driver-class-name: org.postgresql.Driver
+      url: jdbc:postgresql://worlddb:5432/worlddb?allowMultiQueries=true
+      username: hoge
       password: hoge
-    ## userDB
-    user:
-      driver-class-name: com.mysql.cj.jdbc.Driver
-      url: jdbc:mysql://127.0.0.1:3305/user
-      username: root
+    ## usersDB
+    users:
+      driver-class-name: org.postgresql.Driver
+      url: jdbc:postgresql://userdb:5432/usersdb
+      username: hoge
       password: hoge
 ## ログレベルの指定
 logging:
