--- src/main/resources/com/example/demo/repository/mapper/user/UserDbMapper.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE
-	mapper
-	PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-	"http://mybatis.org/dtd/mybatis-3-mapper.dtd"
->
-<mapper namespace="com.example.demo.repository.mapper.user.UserDbMapper">
-	<select
-		id="selectUserCountryCode"
-		parameterType="com.example.demo.repository.entity.user.ReqSelectUserCountryCodeEntity"
-		resultType="com.example.demo.repository.entity.user.ResSelectUserCountryCodeEntity"
-	>
-		SELECT
-			DISTINCT Name, CountryCode
-		FROM
-			users
-		WHERE
-			<foreach
-				item="name" index="index" collection="listName"
-				open="Name IN (" separator=", " close=")"
-			>
-				'${name}'
-			</foreach>
-		ORDER BY
-			Name
-		ASC,
-			CountryCode
-		ASC
-	</select>
-</mapper>
