+++ src/main/resources/com/example/demo/repository/mapper/users/UsersDbMapper.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE
+	mapper
+	PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+	"http://mybatis.org/dtd/mybatis-3-mapper.dtd"
+>
+<mapper namespace="com.example.demo.repository.mapper.users.UsersDbMapper">
+	<select
+		id="selectUserCountryCode"
+		parameterType="com.example.demo.repository.entity.users.ReqSelectUserCountryCodeEntity"
+		resultType="com.example.demo.repository.entity.users.ResSelectUserCountryCodeEntity"
+	>
+		SELECT
+			DISTINCT Name, CountryCode
+		FROM
+			userschema.users
+		WHERE
+			<foreach
+				item="name" index="index" collection="listName"
+				open="Name IN (" separator=", " close=")"
+			>
+				'${name}'
+			</foreach>
+		ORDER BY
+			Name
+		ASC,
+			CountryCode
+		ASC
+	</select>
+</mapper>
